Don't overwrite existing files

+++ script.py
@@ -30,39 +30,46 @@ def ensure_dir(dirname):
         os.makedirs(dirname)
 
 
-def archive(infile, archivedir):
-    ensure_dir(archivedir)
+def archive(infile, archivedir, force=False):
     basename = os.path.basename(infile)
     archivefile = os.path.join(archivedir, basename)
+    if os.path.exists(archivedir) and not force:
+        click.echo("  original couldn't be archived ({} already exists)".format(archivedir))
+        return
+    ensure_dir(archivedir)
     os.rename(infile, archivefile)
     click.echo("  original archived as {}".format(archivefile))
 
 
-def process(infile, archivedir=None, outdir=None):
+def process(infile, force=False, archivedir=None, outdir=None):
     dirname, basename = os.path.split(infile)
+    filename, _ = os.path.splitext(basename)
     if outdir:
         ensure_dir(outdir)
         dirname = outdir
-    filename, _ = os.path.splitext(basename)
     outfile = os.path.join(dirname, filename + '.jpg')
+    if os.path.exists(outfile) and not force:
+        click.echo("  {} skipped! ({} already exists)".format(infile, outfile))
+        return
     click.echo("  {} -> {}".format(infile, outfile))
     img = Image.open(infile)
     img = crop_frame(img)
     img = resize(img, scale=4)
     img.save(outfile, format='JPEG')
     if archivedir:
-        archive(infile, archivedir)
+        archive(infile, archivedir, force=force)
 
 
 @click.command()
 @click.argument('path', type=click.Path(exists=True))
+@click.option('--force', '-f', is_flag=True)
 @click.option('--archivedir', '-a', default=None)
 @click.option('--outdir', '-o', default=None)
-def cli(path, archivedir, outdir):
+def cli(path, force, archivedir, outdir):
     if not os.path.isdir(path):
-        process(path, archivedir=archivedir, outdir=outdir)
+        process(path, force=force, archivedir=archivedir, outdir=outdir)
     else:
         filenames = glob('{}/*.BMP'.format(path))
         click.echo("{} GameBoy Camera pics found!".format(len(filenames)))
         for filename in filenames:
-            process(filename, archivedir=archivedir, outdir=outdir)
+            process(filename, force=force, archivedir=archivedir, outdir=outdir)
